refactor: Improve code structure in JavaScript file
