+++ frontends/scripts/sortable.js
@@ -1,0 +1,43 @@
+import Sortable from "sortablejs";
+
+document.addEventListener("DOMContentLoaded", function () {
+  var element = document.getElementById("faq-list");
+  if (element) {
+    var sortable = Sortable.create(element, {
+      animation: 150,
+      onEnd: function () {
+        var order = sortable.toArray();
+        fetch("/faqs/updated-faq-position/", {
+          method: "POST",
+          headers: {
+            "Content-Type": "application/json",
+            "X-CSRFToken": getCookie("csrftoken"),
+          },
+          body: JSON.stringify({ position: order }),
+        })
+          .then((response) => response.json())
+          .then((data) => {
+            console.log("Success:", data);
+          })
+          .catch((error) => {
+            console.error("Error:", error);
+          });
+      },
+    });
+  }
+});
+
+function getCookie(name) {
+  let cookieValue = null;
+  if (document.cookie && document.cookie !== "") {
+    const cookies = document.cookie.split(";");
+    for (let i = 0; i < cookies.length; i++) {
+      const cookie = cookies[i].trim();
+      if (cookie.substring(0, name.length + 1) === name + "=") {
+        cookieValue = decodeURIComponent(cookie.substring(name.length + 1));
+        break;
+      }
+    }
+  }
+  return cookieValue;
+}
