+++ projects/views.py
@@ -175,17 +175,17 @@ def gender_proportion(request, slug):
     from .models import Sponsor
     project = get_object_or_404(Project, slug=slug)
 
-    # 獲取贊助者的性別數據並分組統計
+    
     gender_data = (
         Sponsor.objects.filter(project=project)  
         .values("account__profile__gender")  
         .annotate(count=Count("id"))  
     )
 
-    # 構建數據
+    
     labels = []
     data = []
-    total_count = 0  # 初始化總人數
+    total_count = 0  
 
     for entry in gender_data:
         gender = entry["account__profile__gender"]
@@ -199,14 +199,14 @@ def gender_proportion(request, slug):
             labels.append("未知")
         count = entry["count"]
         data.append(count)
-        total_count += count  # 累加總人數
+        total_count += count  
 
-    # 返回的 JSON 數據
+    
     response_data = {
         "labels": labels,
         "datasets": [
             {
-                "label": f"贊助者性別比例（總人數: {total_count}）",  # 添加總人數
+                "label": f"贊助者性別比例（總人數: {total_count}）",  
                 "backgroundColor": ["#F8AFAF", "#FFE69B", "#A8D3F0", "#CACACA"],
                 "borderColor": "#FFFFFF",
                 "borderWidth": 2,
@@ -220,10 +220,7 @@ def gender_proportion(request, slug):
 @login_required
 def daily_sponsorship_amount(request, slug):
     from .models import Sponsor
-
     project = get_object_or_404(Project, slug=slug)
-
-    # 檢查是否有贊助紀錄
     sponsors = Sponsor.objects.filter(project=project)
     if not sponsors.exists():
         return JsonResponse({
@@ -237,29 +234,28 @@ def daily_sponsorship_amount(request, slug):
             }]
         })
 
-    # 獲取日期範圍
+    
     start_date = sponsors.order_by("created_at").first().created_at.date()
     end_date = sponsors.order_by("-created_at").first().created_at.date()
 
-    # 用字典記錄每天的實際贊助金額
+    
     daily_totals = {}
     
-    # 先計算每天的贊助總額
+    
     for sponsor in sponsors:
         date = sponsor.created_at.date()
         if date not in daily_totals:
             daily_totals[date] = 0
         daily_totals[date] += sponsor.amount
 
-    # 產生日期序列並計算累積金額
+    
     date_range = []
     cumulative_amount = []
     cumulative_total = 0
     
     current_date = start_date
     while current_date <= end_date:
-        date_range.append(current_date)
-        # 取得當天的贊助金額（如果沒有則為0）並加入累積總額
+        date_range.append(current_date)        
         daily_amount = daily_totals.get(current_date, 0)
         cumulative_total += daily_amount
         cumulative_amount.append(cumulative_total)
@@ -284,14 +280,14 @@ def gender_amount_boxplot(request, slug):
     
     project = get_object_or_404(Project, slug=slug)
 
-    # 過濾贊助數據並分性別進行統計
+    
     gender_data = (
         Sponsor.objects.filter(project=project, reward__isnull=True)
         .select_related('account__profile')
         .values("account__profile__gender", "amount")
     )
 
-    # 構建數據
+    
     gender_amounts = {"M": [], "F": [], "O": []}
 
     for entry in gender_data:
@@ -300,30 +296,30 @@ def gender_amount_boxplot(request, slug):
         if gender in gender_amounts:
             gender_amounts[gender].append(amount)
 
-    # 計算每個性別的箱型圖數據
+    
     boxplot_data = []
     for gender, amounts in gender_amounts.items():
-        if amounts:  # 確保有數據才計算
+        if amounts:  
             sorted_amounts = sorted(amounts)
             n = len(sorted_amounts)
             
-            # 計算四分位數
+            
             q1_idx = int(n * 0.25)
-            q2_idx = int(n * 0.5)  # 中位數
+            q2_idx = int(n * 0.5)  
             q3_idx = int(n * 0.75)
             
             q1 = sorted_amounts[q1_idx]
             median = sorted_amounts[q2_idx]
             q3 = sorted_amounts[q3_idx]
             
-            # 計算四分位距 (使用 Decimal)
+            
             iqr = q3 - q1
             
-            # 計算上下限 (使用 Decimal)
+            
             lower_bound = max(min(sorted_amounts), q1 - iqr * Decimal('1.5'))
             upper_bound = min(max(sorted_amounts), q3 + iqr * Decimal('1.5'))
             
-            # 找出離群值
+            
             outliers = [x for x in sorted_amounts if x < lower_bound or x > upper_bound]
             
             boxplot_data.append({
@@ -365,7 +361,7 @@ def reward_grouped_bar_chart(request, slug):
 
     def get_age_group(birthday):
         if not birthday:
-            return None  # 如果生日不存在，返回 None
+            return None  
         age = (date.today() - birthday).days // 365
         if age < 18:
             return "未成年"
@@ -378,30 +374,30 @@ def reward_grouped_bar_chart(request, slug):
         else:
             return "46+"
 
-    # 修改查詢以排除 reward 為 null 的記錄
+    
     sponsor_data = (
         Sponsor.objects.filter(
             project=project,
-            reward__isnull=False  # 排除 reward 為 null 的記錄
+            reward__isnull=False  
         )
         .select_related("reward", "account__profile")
         .values("reward__title", "account__profile__gender", "account__profile__birthday")
         .annotate(count=Count("id"))
-        .order_by("reward__title")  # 確保結果有序
+        .order_by("reward__title")  
     )
 
-    # 初始化所有可能的組合
+    
     gender_groups = ["M", "F", "O"]
     age_groups = ["未成年", "18-25", "26-35", "36-45", "46+"]
     
-    # 構建數據
+    
     grouped_data = {}
     for entry in sponsor_data:
         reward_title = entry["reward__title"] or "無回饋"
         gender = entry["account__profile__gender"]
         age_group = get_age_group(entry["account__profile__birthday"])
 
-        # 過濾掉 "未知" 性別和年齡組的數據
+        
         if not gender or not age_group:
             continue
 
@@ -417,34 +413,34 @@ def reward_grouped_bar_chart(request, slug):
             
         grouped_data[reward_title][group_key] = count
 
-    # 準備所有可能的組合
+    
     all_combinations = [
         f"{gender}_{age}" 
         for gender in gender_groups 
         for age in age_groups
     ]
 
-    # 構建每個組合的數據集
+    
     labels = list(grouped_data.keys())
     datasets = []
     
-    # 定義固定的顏色映射
+    
     colors = {
-        "M_18-25": "#4E79A7",  # 藍色
-        "M_26-35": "#76B7B2",  # 淺藍
-        "M_36-45": "#59A14F",  # 綠色
-        "M_46+": "#8CD17D",    # 淺綠
-        "M_未成年": "#A0CBE8",  # 淺藍灰
-        "F_18-25": "#F28E2B",  # 橘色
-        "F_26-35": "#FF9DA7",  # 粉紅
-        "F_36-45": "#E15759",  # 紅色
-        "F_46+": "#FFB4A2",    # 淺粉
-        "F_未成年": "#FFB4A2",  # 淺粉
-        "O_18-25": "#9C755F",  # 棕色
-        "O_26-35": "#BAB0AC",  # 灰色
-        "O_36-45": "#808080",  # 深灰
-        "O_46+": "#666666",    # 暗灰
-        "O_未成年": "#CCCCCC",  # 淺灰
+        "M_18-25": "#4E79A7",  
+        "M_26-35": "#76B7B2",  
+        "M_36-45": "#59A14F",  
+        "M_46+": "#8CD17D",    
+        "M_未成年": "#A0CBE8",  
+        "F_18-25": "#F28E2B",  
+        "F_26-35": "#FF9DA7",  
+        "F_36-45": "#E15759",  
+        "F_46+": "#FFB4A2",    
+        "F_未成年": "#FFB4A2",  
+        "O_18-25": "#9C755F",  
+        "O_26-35": "#BAB0AC",  
+        "O_36-45": "#808080",  
+        "O_46+": "#666666",    
+        "O_未成年": "#CCCCCC",  
     }
 
     for combination in all_combinations:
@@ -458,7 +454,6 @@ def reward_grouped_bar_chart(request, slug):
                 for reward in labels
             ]
         }
-        # 只加入有數據的數據集
         if any(dataset["data"]):
             datasets.append(dataset)
